backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
pytesseract
pdf2image
pypdf
python-docx
nltk
scikit-learn
numpy
spacy
textblob
opencv-python==4.8.1.78
